requirements.txt
python-dotenv==1.2.1
newsapi-python==0.2.7
pytest==9.0.2
beautifulsoup4==4.14.3
requests